Extracted test/testing_tools.h
